rename testimonial source cmponent for consistency
